LayoutTests/media/track/track-mode.html

<!DOCTYPE html>
<html>
    <head>

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>

            var textTrack;
            var cueCount;

            function setMode(value)
            {
                run("textTrack.mode = '" + value + "'");
                testExpected("textTrack.mode", value);
                testExpected("video.textTracks[0].mode", value);
                if (value == TextTrack.DISABLED)
                    testExpected("textTrack.cues", null);
                else
                    testExpected("textTrack.cues.length", textTrack.cues.length);
                consoleWrite("");
            }

            function start()
            {
                findMediaElement();
                cueCount = 0;
                consoleWrite("");

                textTrack = document.getElementById('track_1').track;

                consoleWrite("<b>++ Test default attribute value</b>");
                testExpected("textTrack.mode", TextTrack.SHOWING);
                testExpected("video.textTracks[0].mode", TextTrack.SHOWING);
                consoleWrite("");

                consoleWrite("<b>*** Set to bogus value, should return default</b>");
                var value = 37;
                testException("textTrack.mode = '" + value + "'", "DOMException.INVALID_ACCESS_ERR");
                testExpected("textTrack.mode", TextTrack.SHOWING);
                testExpected("video.textTracks[0].mode", TextTrack.SHOWING);
                consoleWrite("");

                consoleWrite("<b>*** Set to known values</b>");
                consoleWrite("<b>++ 'TextTrack.DISABLED'</b>");
                setMode(TextTrack.DISABLED);

                video.src = findMediaFile("video", "../content/test");
                video.play();
                // Wait for end of first cue (no events should fire while track is disabled).
                setTimeout(testHiddenAndShowing, 400);
                consoleWrite("++ 0 events expected while mode = TextTrack.DISABLED");
                consoleWrite("");
            }
            
            function testHiddenAndShowing()
            {
                consoleWrite("<b>++ 'TextTrack.HIDDEN'</b>");
                setMode(TextTrack.HIDDEN);

                consoleWrite("<b>++ 'TextTrack.SHOWING'</b>");
                setMode(TextTrack.SHOWING);

                consoleWrite("++ at least " + textTrack.cues.length + " events expected while mode = TextTrack.SHOWING");
            }
            
            waitForEvent('cuechange',
                function (event)
                {
                    cueCount++;
                    if (cueCount == textTrack.cues.length)
                        endTest();

                }
            );
            
        </script>
    </head>
    <body>
        <p>Tests that the TextTrack mode attribute is appropriately set.</p>
        <video>
            <track id=track_1 src="captions-webvtt/captions-fast.vtt" onload="start()" default>
        </video>
    </body>
</html>
